fix(DTFS2-6425): resolved comments on PR
